requirements.txt
pynput==1.7.6
pywin32==306
python-dotenv==1.0.0
pystray
Pillow
